start com/handle-ws-inbound :save-response

+++ resources/migrations/20190506000098-alter-round_product-table.up.sql
@@ -1,3 +1,3 @@
 ALTER TABLE vetd.round_product 
-ADD COLUMN result bigint,
+ADD COLUMN result integer,
 ADD COLUMN reason text
